Basic questing system, with sample 'testevent' and 'testlistener'
Events and Listeners are generics - write listeners as ListenerSample : QuestListener<T, U> and events as EventSample : QuestEvent<U, T> with T being ListenerSample and U being EventSample. Listeners have a virtual awake, override it BUT do not have base.Awake() ANYWHERE but at the start of the function. Events require a a constructor that calls the base constructor (make sure the bool is set to true), as well as takes in any parameters it might need to save to parrot back to the listeners.